requirements.txt
alembic==1.7.7
aniso8601==9.0.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.6
dataclasses-json==0.5.7
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
psycopg2==2.7.7
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.7.0
